+++ R/package.R
@@ -68,11 +68,21 @@ keras <- NULL
     "ipython" #, "tensorflow_datasets"
   ))
 
+  if (is.na(Sys.getenv("KERAS_HOME", NA))) {
+    if (!dir.exists("~/.keras/")) {
+      Sys.setenv("KERAS_HOME" = normalizePath(
+        tools::R_user_dir("keras3", "cache"),
+        mustWork = FALSE
+        ))
+    }
+  }
+
   # default backend is tensorflow for now
   # the tensorflow R package calls `py_require()` to ensure GPU is usable on Linux
   # use_backend() includes py_require(action = "remove") calls to undo
   # what tensorflow:::.onLoad() did. Keep them in sync!
   # backend <- Sys.getenv("KERAS_BACKEND", "jax")
+  # ~/.keras.keras.json also has an (undocumented) 'backend' field
   backend <- Sys.getenv("KERAS_BACKEND", "tensorflow")
   gpu <- NA
   if (endsWith(backend, "-cpu")) {
@@ -101,11 +111,12 @@ keras <- NULL
       # check version
       # check_implementation_version()
 
-      tryCatch(
-        import("tensorflow")$experimental$numpy$experimental_enable_numpy_behavior(),
-        error = function(e) {
-          warning("failed setting experimental_enable_numpy_behavior")
-        })
+      # disabled because of errors with keras-hub
+      # tryCatch(
+      #   import("tensorflow")$experimental$numpy$experimental_enable_numpy_behavior(),
+      #   error = function(e) {
+      #     warning("failed setting experimental_enable_numpy_behavior")
+      #   })
 
     },